Cleanup: remove dead code and consolidate options
